refactor(package-manager): move package registry into dedicated app
