+++ apps/package_manager/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class PackageManagerConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "apps.package_manager"
