+++ apps/package_manager/tests/test_package_registry_api.py
